requirements.txt
pandas
langchain_openai 
langchain_core
langchain
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
groq
cassio
beautifulsoup4
langchain-groq
wikipedia
arxiv
langchainhub
nltk
PyPDF2
google-cloud-aiplatform>=1.38
langchain_google_genai
ragas
chardet
plotly
pydantic
typing
langchain_experimental
tabulate
streamlit-autorefresh
beautifultable
camelot-py
opencv-python
ghostscript
lxml
tenacity
nltk
matplotlib
scipy
langchain-text-splitters